Use BigInteger for score type

+++ migration/versions/5f31dcc0e896_score_type_integer.py
@@ -16,7 +16,7 @@ down_revision = "3e307dd3c114"
 
 def upgrade():
     with op.batch_alter_table("output") as batch_op:
-        batch_op.alter_column("score", type_=sa.Integer)
+        batch_op.alter_column("score", type_=sa.BigInteger)
 
 
 def downgrade():
